requirements.txt
redis==3.5.3
pymongo==3.11.3
mysql-connector-python==8.0.24
